Update branch name to #master
When I forked this code, it was created as #master (a branch named master) instead of main. This caused an error of 'Error: failed to download context. Please check if the URL is incorrect. If it has credentials, please check if they are expired' then I ran az acr task run --registry $ACR_NAME --name taskhelloworld  . Propose to rename all instances of #main with #master